removed gnome-mplayer, added mpv
mpv is WAY better. gnome-mplayer (and anything mplayer based) won't scale. gstreamer based programs are inferior such as parole, which won't fullscreen smaller videos and has worse performance. optionally you could use smplayer and smtube, but that's extra. mpv optionally could use youtube-dl.

not sure of any codecs or dependencies for x264, x265, and other video. ffmpeg is installed due to firefox dependency.
mpv performance for x264 is excellent, and x265 is almost perfect.
playing 1080p videos is not recommended- the screen is only 720p (plus a few extra pixels at 1366x768), and performance is not watchable for 1080p videos that have to be scaled back to fit the screen.